# meta/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Sportfest backend",
    "version": "1.0.0",
    "title": "Sportfest Backend",
    "contact": {
      "url": "https://github.com/atiw-sportfest"
    }
  },
  "host": "sportfest.atiw.de",
  "basePath": "/backend",
...
  "schemes": [
    "https"
  ],
  "parameters": {
...
    "did": {
      "in": "path",
      "name": "did",
      "description": "Disziplin-ID",
      "required": true,
      "type": "integer",
      "format": "int64"
    },
...
    "sid": {
      "in": "path",
      "name": "sid",
      "description": "Schueler-ID",
      "required": true,
      "type": "integer",
      "format": "int64"
    },
...
  },
  "responses": {
...
    "Anmeldungen": {
      "description": "Anmeldungen",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Anmeldung"
        }
      }
    },
...
    "Ergebnisse": {
      "description": "Ergebnisse",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Ergebnis"
        }
      }
    },
...
    "Schueler": {
      "description": "Schueler",
      "schema": {
        "$ref": "#/definitions/Schueler"
      }
    },
...
  },
  "paths": {
...
    "/schueler/{sid}/": {
      "get": {
        "summary": "Schueler abrufen",
        "tags": [
          "Teilnehmer"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/sid"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Schueler"
          }
        }
      },
      "delete": {
        "summary": "Schueler löschen",
        "tags": [
          "Teilnehmer"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/sid"
          }
        ],
        "responses": {
          "204": {
            "description": "Schueler gelöscht."
          }
        }
      }
    },
    "/schueler/{sid}/anmeldungen": {
      "get": {
        "summary": "Anmeldungen eines Schuelers anzuzeigen",
        "tags": [
          "Anmeldung"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/sid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Anmeldungen"
          }
        }
      }
    },
    "/schueler/{sid}/ergebnisse": {
      "get": {
        "summary": "Ergebnisse eines Schuelers anzuzeigen",
        "tags": [
          "Teilnehmer"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/sid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Ergebnisse"
          }
        }
      }
    },
    "/schueler/{sid}/ergebnisse/{did}": {
      "get": {
        "summary": "Ergebnisse eines Schuelers fuer eine Disziplin anzeigen",
        "tags": [
          "Teilnehmer"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/did"
          },
          {
            "$ref": "#/parameters/sid"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Ergebnisse"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "jwt": {
      "description": "JWT Bearer-Authentication",
      "type": "apiKey",
      "name": "HTTP-Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Anmeldung": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "disziplin": {
          "$ref": "#/definitions/Disziplin"
        },
        "schueler": {
          "$ref": "#/definitions/Schueler"
        }
      }
    },
    "Disziplin": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "bezeichnung": {
          "type": "string"
        },
        "beschreibung": {
          "type": "string"
        },
        "team": {
          "type": "boolean"
        },
        "versus": {
          "type": "boolean"
        },
        "variablen": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Variable"
          }
        },
        "regeln": {
          "type": "string"
        }
      }
    },
    "Ergebnis": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "punkte": {
          "type": "integer"
        },
        "leistungen": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Leistung"
          }
        },
        "klasse": {
          "$ref": "#/definitions/Klasse"
        },
        "schueler": {
          "$ref": "#/definitions/Schueler"
        },
        "disziplin": {
          "$ref": "#/definitions/Disziplin"
        }
      }
    },
    "Klasse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "bezeichnung": {
          "type": "string"
        }
      }
    },
    "Leistung": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "wert": {
          "type": "string"
        },
        "variable": {
          "$ref": "#/definitions/Variable"
        }
      }
    },
    "Schueler": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "vorname": {
          "type": "string"
        },
        "nachname": {
          "type": "string"
        },
        "geschlecht": {
          "type": "string"
        },
        "klasse": {
          "$ref": "#/definitions/Klasse"
        }
      }
    },
    "Typ": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "datentyp": {
          "type": "string",
          "enum": [
            "string",
            "int",
            "float",
            "long",
            "double"
          ]
        },
        "einheit": {
          "type": "string"
        },
        "format": {
          "type": "string",
          "description": "RegEx"
        },
        "bsp": {
          "type": "string",
          "description": "Beispiel"
        }
      }
    },
    "Variable": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "bezeichnung": {
          "type": "string"
        },
        "typ": {
          "$ref": "#/definitions/Typ"
        }
      }
    },
...
  }
}